requirements.txt
altgraph==0.17
attrs==19.3.0
auto-py-to-exe==2.7.3
autopep8==1.5.2
backcall==0.1.0
bleach==3.1.5
bottle==0.12.18
bottle-websocket==0.2.9
decorator==4.4.2
defusedxml==0.6.0
distro==1.5.0
Eel==0.11.0
entrypoints==0.3
future==0.18.2
gevent==20.5.2
gevent-websocket==0.10.1
greenlet==0.4.15
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.0
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
MarkupSafe==1.1.1
mistune==0.8.4
mypy==0.770
mypy-extensions==0.4.3
nbconvert==5.6.1
nbformat==5.0.6
notebook==6.0.3
numpy==1.18.4
packaging==20.4
pandas==1.0.4
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.8.0
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pycodestyle==2.6.0
Pygments==2.6.1
PyInstaller==3.6
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
pyzmq==19.0.1
qtconsole==4.7.4
QtPy==1.9.0
Send2Trash==1.5.0
six==1.15.0
tabula-py==2.1.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions==3.7.4.2
wcwidth==0.2.3
webencodings==0.5.1
whichcraft==0.6.1
widgetsnbextension==3.5.1
zope.event==4.4
zope.interface==5.1.0
